feat: add resend library (#106)
* feat: add resend library

* feat: create send route email

* feat: create nex email template

* feat: update auth action with resend

* fix: fix typo send email url
