+++ package.json
@@ -22,6 +22,7 @@
     "@types/nodemailer": "^6.4.17",
     "@vercel/blob": "^0.27.1",
     "@web-std/file": "^3.0.3",
+    "googleapis": "^146.0.0",
     "jose": "^5.9.6",
     "next": "14.1.0",
     "next-intl": "^3.13.0",
@@ -32,6 +33,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-responsive-carousel": "^3.2.23",
+    "resend": "^4.1.2",
     "sass": "^1.70.0",
     "ts-node": "^10.9.2"
   },
